Create M3K4script.m
1 - I noticed is that the masses are not staying in their locations.
Does not matter for now, but going forward worth noting that the
initial conditions (K values, m values and initial position/ velocity
values) need to be considered thoroughly or else it won't be real
world relevant.

2 - Try and keep time pretty short because otherwise graph gets so
noisy it is hard to try and see what is going on.

3 - You need to have both files M3K4script and M3K4, which is a
function. you must save both to same area of matlab. To run the code
you run the script, which will call the function.